requirements.txt
numpy
opencv-python
torch
torchvision
detectron2 @ https://dl.fbaipublicfiles.com/detectron2/wheels/cu118/torch2.0/detectron2-0.6-cp310-cp310-linux_x86_64.whl
pycocotools
albumentations

# NLP
transformers
datasets
sentence-transformers
faiss-cpu
pandas
seqeval
jupyterlab